renamed model.py to deployment_client.py
